refactor: move converter from utils
Move `linkml.utils.converter` to `linkml.converter.cli`, which is more
consistent with the structuring of other tools in the repo:
- Package for the top-level functionality.
- CLI specific to that functionality inside of that package.
- Integration into higher-level CLI.
