+++ pyproject.toml
@@ -128,7 +128,7 @@ gen-sqltables = "linkml.generators.sqltablegen:cli"
 gen-summary = "linkml.generators.summarygen:cli"
 gen-project = "linkml.generators.projectgen:cli"
 run-tutorial = "linkml.utils.execute_tutorial:cli"
-linkml-convert = "linkml.utils.converter:cli"
+linkml-convert = "linkml.converter.cli:cli"
 linkml-lint = "linkml.linter.cli:main"
 linkml-sqldb = "linkml.utils.sqlutils:main"
 linkml-validate = "linkml.validator.cli:cli"
